Update calculator.c
a double

+++ src/lib/calculator.c
@@ -16,4 +16,4 @@ double _div(double a, double b) {
 	return a/b;
 }
 
-double car(int a) { return a * a; }
+double car(double a) { return a * a; }
